requirements.txt
langchain
unstructured
pydantic
python-dotenv
tqdm
loguru
PyPDF2
tiktoken

# Dev / test tools
pytest
pytest-cov
black
flake8
mypy
